Add constant PermissionTypes
