+++ src/constants/PermissionTypes.ts
@@ -1,0 +1,14 @@
+const PermissionTypes = {
+  AANBOD_BEWERKEN: 'AANBOD_BEWERKEN',
+  AANBOD_MODEREREN: 'AANBOD_MODEREREN',
+  AANBOD_VERWIJDEREN: 'AANBOD_VERWIJDEREN',
+  ORGANISATIES_BEWERKEN: 'ORGANISATIES_BEWERKEN',
+  ORGANISATIES_BEHEREN: 'ORGANISATIES_BEHEREN',
+  GEBRUIKERS_BEHEREN: 'GEBRUIKERS_BEHEREN',
+  LABELS_BEHEREN: 'LABELS_BEHEREN',
+  VOORZIENINGEN_BEWERKEN: 'VOORZIENINGEN_BEWERKEN',
+  PRODUCTIES_AANMAKEN: 'PRODUCTIES_AANMAKEN',
+  FILMS_AANMAKEN: 'FILMS_AANMAKEN',
+} as const;
+
+export { PermissionTypes };
